+++ src/RadixDlt.NetworkGateway.PostgresIntegration/LedgerExtension/ReadHelper.cs
@@ -441,7 +441,7 @@ SELECT
     nextval('account_authorized_depositor_aggregate_history_id_seq') AS AccountAuthorizedDepositorAggregateHistorySequence,
     nextval('unverified_standard_metadata_aggregate_history_id_seq') AS UnverifiedStandardMetadataAggregateHistorySequence,
     nextval('unverified_standard_metadata_entry_history_id_seq') AS UnverifiedStandardMetadataEntryHistorySequence,
-    nextval('resource_owners_id_seq') AS ResourceOwnersSequence
+    nextval('resource_holders_id_seq') AS ResourceHoldersSequence
 ",
             cancellationToken: token);