+++ src/RadixDlt.NetworkGateway.PostgresIntegration/LedgerExtension/WriteHelper.cs
@@ -735,7 +735,7 @@ internal class WriteHelper : IWriteHelper
         return entities.Count;
     }
 
-    public async Task<int> CopyResourceOwners(ICollection<ResourceOwners> entities, CancellationToken token)
+    public async Task<int> CopyResourceHolders(ICollection<ResourceHolder> entities, CancellationToken token)
     {
         if (!entities.Any())
         {
@@ -746,15 +746,15 @@ internal class WriteHelper : IWriteHelper
 
         await using var createTempTableCommand = _connection.CreateCommand();
         createTempTableCommand.CommandText = @"
-CREATE TEMP TABLE tmp_resource_owners
-(LIKE resource_owners INCLUDING DEFAULTS)
+CREATE TEMP TABLE tmp_resource_holders
+(LIKE resource_holders INCLUDING DEFAULTS)
 ON COMMIT DROP";
 
         await createTempTableCommand.ExecuteNonQueryAsync(token);
 
         await using var writer =
             await _connection.BeginBinaryImportAsync(
-                "COPY tmp_resource_owners (id, entity_id, resource_entity_id, balance, last_updated_at_state_version) FROM STDIN (FORMAT BINARY)",
+                "COPY tmp_resource_holders (id, entity_id, resource_entity_id, balance, last_updated_at_state_version) FROM STDIN (FORMAT BINARY)",
                 token);
 
         foreach (var e in entities)
@@ -772,8 +772,8 @@ ON COMMIT DROP";
 
         await using var mergeCommand = _connection.CreateCommand();
         mergeCommand.CommandText = @"
-MERGE INTO resource_owners ro
-USING tmp_resource_owners tmp
+MERGE INTO resource_holders ro
+USING tmp_resource_holders tmp
 ON ro.entity_id = tmp.entity_id AND ro.resource_entity_id = tmp.resource_entity_id
 WHEN MATCHED AND tmp.balance = 0 THEN DELETE
 WHEN MATCHED AND tmp.balance != 0 THEN UPDATE SET balance = tmp.balance, last_updated_at_state_version = tmp.last_updated_at_state_version
@@ -781,7 +781,7 @@ WHEN NOT MATCHED AND tmp.balance != 0 THEN INSERT VALUES(id, entity_id, resource
 
         await mergeCommand.ExecuteNonQueryAsync(token);
 
-        await _observers.ForEachAsync(x => x.StageCompleted(nameof(CopyResourceOwners), Stopwatch.GetElapsedTime(sw), entities.Count));
+        await _observers.ForEachAsync(x => x.StageCompleted(nameof(CopyResourceHolders), Stopwatch.GetElapsedTime(sw), entities.Count));
 
         return entities.Count;
     }
@@ -834,7 +834,7 @@ SELECT
     setval('account_authorized_depositor_aggregate_history_id_seq', @accountAuthorizedDepositorAggregateHistorySequence),
     setval('unverified_standard_metadata_aggregate_history_id_seq', @unverifiedStandardMetadataAggregateHistorySequence),
     setval('unverified_standard_metadata_entry_history_id_seq', @unverifiedStandardMetadataEntryHistorySequence),
-    setval('resource_owners_id_seq', @resourceOwnersSequence)
+    setval('resource_holders_id_seq', @resourceHoldersSequence)
 ",
             parameters: new
             {
@@ -879,7 +879,7 @@ SELECT
                 accountAuthorizedDepositorAggregateHistorySequence = sequences.AccountAuthorizedDepositorAggregateHistorySequence,
                 unverifiedStandardMetadataAggregateHistorySequence = sequences.UnverifiedStandardMetadataAggregateHistorySequence,
                 unverifiedStandardMetadataEntryHistorySequence = sequences.UnverifiedStandardMetadataEntryHistorySequence,
-                resourceOwnersSequence = sequences.ResourceOwnersSequence,
+                resourceHoldersSequence = sequences.ResourceHoldersSequence,
             },
             cancellationToken: token);