+++ src/RadixDlt.NetworkGateway.PostgresIntegration/Migrations/20240816104539_InitialCreate.Designer.cs
@@ -81,7 +81,7 @@ using RadixDlt.NetworkGateway.PostgresIntegration.Models;
 namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
 {
     [DbContext(typeof(MigrationsDbContext))]
-    [Migration("20240815152003_InitialCreate")]
+    [Migration("20240816104539_InitialCreate")]
     partial class InitialCreate
     {
         /// <inheritdoc />
@@ -1557,7 +1557,7 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
                     b.ToTable("resource_entity_supply_history");
                 });
 
-            modelBuilder.Entity("RadixDlt.NetworkGateway.PostgresIntegration.Models.ResourceOwners", b =>
+            modelBuilder.Entity("RadixDlt.NetworkGateway.PostgresIntegration.Models.ResourceHolder", b =>
                 {
                     b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
@@ -1590,7 +1590,7 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
 
                     b.HasIndex("EntityId", "ResourceEntityId", "Balance");
 
-                    b.ToTable("resource_owners");
+                    b.ToTable("resource_holders");
                 });
 
             modelBuilder.Entity("RadixDlt.NetworkGateway.PostgresIntegration.Models.SchemaEntryAggregateHistory", b =>